[Fix] Rename module name to avoid generate route by Docusaurus
